pyproject.toml
[build-system]
requires = [ "poetry>=0.12",]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "remotedata"
version = "0.0.1"
description = "Accessing and caching remote data."
authors = [ "pwwang <pwwang@pwwang.com>",]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/pwwang/remotedata"
repository = "https://github.com/pwwang/remotedata"

[tool.poetry.dependencies]
python = "^3.4"
python-simpleconf = "*"
requests = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
